+++ lib/cloudmunda.rb
@@ -1,8 +1,28 @@
 # frozen_string_literal: true
 
+require 'concurrent'
+
 require_relative "cloudmunda/version"
+require_relative 'cloudmunda/loggable'
+require_relative 'cloudmunda/cli/worker'
+require_relative 'cloudmunda/configuration'
+require_relative 'cloudmunda/api'
+require_relative 'cloudmunda/zeebe'
+require_relative 'cloudmunda/graphql'
 
 module Cloudmunda
-  class Error < StandardError; end
-  # Your code goes here...
+  extend Configuration
+  extend Loggable
+
+  def self.client
+    @client ||= ::Cloudmunda::Zeebe::Client.new
+  end
+
+  def self.register_worker(worker)
+    self.workers << worker
+  end
+
+  def self.workers
+    @workers ||= []
+  end
 end