+++ lib/cloudmunda/api/o_auth_resource.rb
@@ -1,0 +1,19 @@
+# frozen_string_literal: true
+
+require 'ostruct'
+
+module Cloudmunda
+  module API
+    class OAuthResource < OpenStruct
+      def self.create_by_uri(uri:, payload:)
+        raw_item = Cloudmunda::API::Client.post(uri, payload)
+        raw_item = {} if raw_item == ""
+        new(raw_item)
+      end
+
+      def initialize(raw)
+        super(raw)
+      end
+    end
+  end
+end