+++ lib/cloudmunda/graphql.rb
@@ -1,0 +1,2 @@
+require_relative 'graphql/client'
+require_relative 'graphql/user_tasks'