+++ lib/cloudmunda/graphql/client.rb
@@ -1,0 +1,29 @@
+# frozen_string_literal: true
+
+require 'rest-client'
+require 'json'
+require 'uri'
+require 'base64'
+
+module Cloudmunda
+  module Graphql
+    class Client
+      def self.post(params = {})
+        response = RestClient.post(Cloudmunda.graphql_url, params.to_json, headers.merge(content_headers))
+        JSON.parse(response.body)
+      end
+
+      def self.headers
+        {
+          authorization: "Bearer " + Cloudmunda::API::AccessToken.create(audience_url: 'tasklist.camunda.io').access_token
+        }
+      end
+
+      def self.content_headers
+        {
+          'Content-Type': 'application/json'
+        }
+      end
+    end
+  end
+end