+++ lib/cloudmunda/loggable.rb
@@ -1,0 +1,15 @@
+module Cloudmunda
+  module Loggable
+    def logger
+      @logger || setup_logger
+    end
+
+    def logger=(logger)
+      @logger = logger
+    end
+
+    def setup_logger
+      @logger = Beez.config.logger
+    end
+  end
+end