+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 ## [Unreleased]
 
+## [0.1.1] - 2021-12-16
+
+- Camunda Cloud Access Token Creation
+- Zeebe task workers
+- Camunda Cloud Usertasks graphQL connection
+
 ## [0.1.0] - 2021-12-16
 
 - Initial release
+_